requirements.txt
duckduckgo-search
gradio
gradio[oauth]
IPython
langchain_community
langchain_core
langgraph
pyppeteer
requests
boto3
langchain-aws>=0.1.4